web/requirements-base.txt
bs4==0.0.1
django==2.2.13
django-annoying==0.10.4
django-autocomplete-light==3.3.5
django-crispy-forms==1.7.2
django-debug-toolbar==1.11
django-extensions==2.1.9
django-filter==2.1.0
django-jet==1.0.8
django-picklefield ==2.0
django-redis==4.10.0
django-registration-redux==2.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
django-stronghold==0.3.0
django-widget-tweaks==1.4.5
elasticsearch-dsl==7.0.0
gunicorn==19.9.0
pandas==0.24.2
plotly==4.6.0
psutil==5.7.0
psycopg2==2.8.3
redis==3.2.1
requests==2.22.0
scipy==1.3.3
sentry-sdk==0.9.2
xlrd==1.2.0
xlwt==1.3.0
scikit-learn==0.22.2
numpy==1.18.3
